Include output path in file name pattern
When the `lazy` option is set, the `filename` pattern is used to
determine which requests should trigger a reload. Because the string
being tested in that case includes the output path (see the
`getFilenameFromUrl` function), the matching pattern should also include
it.

Update the output path in the `filename` pattern so that the `lazy`
option correctly enables building on request.