Use platform-agnostic pathJoin over path.join
Using path.join for memory-fs leads to problems on Windows, because the
path module will be path.win32 instead of path.unix, which of course uses
backslash as the directory separator. We could explicitly import
path.posix, but since this same JS file already includes a generic
path.join that seems to work in enough cases for us, it seems OK to just
use that.