Support Range requests commonly used by media
Seeking through media, such as audio and video, often leads a browser
to make range requests, in addition to other scenarios. Responding with
the full content instead of the requested range and can lead to
unexpected behaviour in a web application that in production will
otherwise work since the production server supports range requests.

- Depend on range-parser to parse range request headers
- Use source from the `send` npm package on responding to range
  requests https://github.com/pillarjs/send/blob/31ceb6aaf530152c29a1855bf3f9dbd86533303e/index.js#L551-L583