Removing watching/watching.running check
Given that `state` should always have the correct value (if the bundle is valid
or not) then we should be able to use ready directly without checking the
watcher.

I also hit a bug where the watcher had running set to false while the bundle was
invalid, leading to a callback being executed immediately instead of waiting.