Add a fallback argument to handle unreachable file
Imagine, we are debugging a react app with react-router. We click
a link sending us to another url. That's ok since all the routing
stuff are handled in the front end. Then we make some modification
to the html and want to refresh the whole page. Now problem arises,
the url is now unreachable in the back end because we have changed
it so that the server has no idea what file to serve.

In this case, we would like the dev server to serve one similiar
html no matter what the url is( the root html), that woule be
the behaviour if this commit is merged.

The default behaviour is untouched if the fallBack argument is
not true.