Use options.log, warn and error for compilation messages
When the compilation fails or has warnings, the compilation
messages are outputted in console.log not using the options.error
and options.warn functions.

This commit change the defaultReporter to output the stats
and final message to the options.error when the compilation
has errors or the options.warn when it has warnings.

The Reporter.test.js test cases are updated accordingly.