fixes #270: bad filename generation
Uses an existing dependency in the module to test if the generated
filename is absolute before adding the leading slash. This is the most
minimal fix for the Windows incompatibility that doesn't break a
bunch of test cases that currently pass on Windows, and which won't
run afoul of the filesystem abstraction layer used by Webpack.