Attempt to appease the Linux spirits
The pathabs call wasn't treating c:\ as absolute on Linux, so it was
still prepending the /, and hence failing out the test. Now instead
we call both the posix and win32 versions of the function explicitly,
same as is done in the setFs call. That feels dirty to me, but this is
already fairly dirty, and I don't think this opens any kind of path
attack vector that wasn't already present.